services/inflection-analysis/pyproject.toml
[project]
name = "inflection-analysis"
version = "0.1.0"
description = "Voice inflection and emotion analysis service for Verbatim Studio"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "numpy>=1.26.0",
    "scipy>=1.12.0",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "torch>=2.1.0",
    "transformers>=4.36.0",
    "praat-parselmouth>=0.4.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
